requirements.txt
alembic==1.10.3
aniso8601==9.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Deprecated==1.2.13
Flask==2.2.3
Flask-HTTPAuth==4.7.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
gcloud==0.18.3
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.2.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
oauth2client==4.1.3
packaging==23.0
protobuf==4.22.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase4==4.6.0
python-decouple==3.8
python-jwt==4.0.0
pytz==2023.3
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
